Auto-generate schema.h and schema.cpp on compile-time. (#37)
Automatically generate the schema files when running CMake. This
makes sure that these files are always up to date. Plus, we do not
have to maintain them in the repository any longer. And most
importantly, it prevents developer from changing schema.h and
schema.cpp instead of modifying the corresponding files in the schema/
directory.