Fix some gcc warnings. (#41)
* Reorder initializer list to fix gcc warnings.

The members of a class are initialized in the order they appear
in the header file. If the order is different in the initializer
list, it makes gcc trip.

* Fix documentation.

* All Base classes should have a virtual destructor.

Having a virtual destructor is necessary to make sure that the
destructors from both the parent and the child class are called.
The default destructors are not virtual.