Make building tests optional (#46)
* Declare some integers as unsigned.

This fixes some gcc warnings of comparing signed with unsigned
integers. We should take that serious. Comparing a negative integer
to a positive unsigned integer must never happen.

* CMake command enable_testing() should be in the source directory root.

See CMake documentation for details.

* Make building tests (and GoogleTest) optional.

If you want to compile and run the tests, add the following option
to your cmake call: -DTEST=ON