Distinguish between Debug and Release build.
Set the build type in config.h as BUILD_TYPE_DEBUG, BUILD_TYPE_RELEASE,
etc. If the cmake option -DTEST=ON is used, the build type defaults to
"Debug". Otherwise it defaults to "Release".

If the build type is "Debug", the code is compiled with full warnings.
If the build type id "Release", the code is compiled with full
optimization.