Unit Test for UnwrapCoordinates (#109)
The unit test covers particles in the box and outside the box for
a cubic, an orthorombic, a monclinic and a triclinic box.

Implementing the unit test, I have realized that you can actually
give fractional values for the box images. This should be changed
and only take integers (we need a IntVector3 for this).