Unit tests for string methods; general string method fixes (#222)
* Fix StringMethod tolerance check for undefined tol
Changed TolCheck() for the case where a user does not
provide tolerance criteria: If the user does not provide
a vector of tolerance values in the string method inputs,
the vector `_tol` has size 0. TolCheck() needs to return
false in this case, and was previously checking whether
`_tol.size() == 1 && _tol[0] < 0`, which would
never be the case if tolerance values were not provided.
TolCheck() proceeds to do a regular tolerance check
as if tolerance values were provided. This leads to a
seg fault since _tol is empty.

* String unit tests

* Updates to string methods

* Updated CheckEnd in ElasticBand.cpp