Fix for Issue #187 (#239)
* Sync umbrella/initialization across SM walkers

FTS and Swarm would hang indefinitely from the start of the
simulation when run with GROMACS, stemming from different
string walkers being at different timesteps in the simulation
since the number of umbrella (FTS) or initialization (Swarm)
timesteps were allowed to vary for each walker. By ensuring
that the number of umbrella and initialization steps are the
same for all walkers we avoid the GROMACS hang (and this
should still work with other engines e.g. LAMMPS)

* Update functions in StringMethod unit tests

Some function names and parameters were changed in
StringMethod.h, change the StringMethod and Swarm
unit tests to reflect those modifications.

* Update readme to reflect string method fixes

* Fixes #187