Improve Grid access.
The Grid can now be accessed via grid.at(index) or grid[index].
Here, index may be a vector of ints specifying a grid index or a
vector of floats specifying a point. For 1d-grids, index may be a
single int or a single float (instead of a 1-element vector).