Add member functions size() and data()
size() returns the total number of grid points, including under-
and overflow bins. This is equal to the size of the internal storage
vector. data() returns a pointer to the storage vector.

Even though I don't like making the internal workings accessible
from the outside, this is necessary to communicate the data via
MPI.