Add compatibility for GROMACS 2019 and 2018.{4,5}
There are quite a few changes for 2019, as GROMACS is undergoing somewhat of a restructuring.
For the version 2019 series moving forward, we may not be able to use the same code, depending on how much the code base changes.
An added `#include` line in 2018.4 necessitates a second patch file for proper matching.
2018.5 is the same as 2018.4.
The 2016 series has ended, with 2016.5 being its last release.