Fix erroneous line in CMakeLists.txt
The link_directories command requires to specify directories, the
variable passed to it however contains library files. Most linkers
ignore this problem, but we should not rely on this.

The link_directories adds a -L flag to the linker but does not give
a directory but a file. It is not defined how linkers will treat this
(ignore, error, use dirname).